+++ pyproject.toml
@@ -238,6 +238,7 @@ exclude = []
     "PLW0127", # Re-export things for tab-completion explicitly.
     "PT028", # Cannot have default values in test functions with fixtures -- not an issue here.
 ]
+"src/gettsim/plot_dag/__init__.py" = ["PLW0127"] # Re-export things for tab-completion explicitly.
 
 # Mostly things vectorization can't handle
 "src/_gettsim/*" = ["E501", "PLR1714", "PLR1716", "E721", "SIM108", "RET"]
